refactor: extract INFO Config deserializer
